+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `full_moon::Error` and `full_moon::ast::Ast` now implement Serialize and Deserialize.
 - Added optional "stacker" feature which uses the [stacker](https://docs.rs/stacker/latest/stacker/index.html) crate to conditionally expand stack size to avoid stack overflows, a known problem with full-moon.
 - Added support for string interpolation under the `roblox` feature flag
+- Added support for Lua 5.2 fractional hexidecimal / hexidecimal with exponents, and LuaJIT number suffixes (LL/ULL/i) under the `lua52` feature flag
 
 ### Fixed
 - Support instantiated generics with no parameters, e.g. `Foo<>`